Created Keylogger.py
This basic keylogger uses the pynput module utilizing the listening to keyboard functionality. Additionally, the program utilizes SMTP protocol and access a gmail account to send the keystrokes of a targeted device to a specified email. This basic keylogger is the foundation for the spyware that I am developing and is soon coming.